bin: fix the call to the prepare targets for non-selected packages
Right now, we do a normal build for selected packages,
and then call the prepare targets for the remaining
packages so that (among other things) we update the
environment variables.

This call to the prepare target should be done after everything
else, and Autobuild.do_update should be false to avoid unwanted
updates.