fix env_set and env_add w.r.t. latest changes in autobuild
Autobuild now makes the difference between an envvar
that has been set with env_set and one that has not
been changed at all. In the latter case, a call to
env_add will import the current value from ENV.

Autoproj's wrappers on env_set and env_add needed to be changed
to reflect that modification.