add a method that normalizes package selection
Instead of doing it "on the fly", we now preprocess the
user's package selection and return the set of package
names that are actually selected.

Moreover, we now allow directory names as package selection,
and we interpret selection strings as regular expression.
Meaning that, now "modules/" will select everything named
"modules/.*" for instance.