fix the behaviour of Manifest#each_source when the block calls break
The problem is that we now cache the sources, and we were yielding
them at the same time that the @sources array was being built.
Issue: if break is called in yield(source), then the @sources array
will end up being partial *and* will never be updated.

So, we now first build @sources and only then enumerate it