first try at fixing the handling of sublayouts
The sublayouts are now flattened: the import and build is done in
a way that makes the layout boundaries transparent and should
result in a better overall behaviour.

User-level change: there is now only one env.sh script. As a
backward compatibility measure, the other env.sh scripts now
source it.

Bug left to fix: how the dependency discovery is done results in
having the orogen dependencies *not* detected if packages are
selected on the command line.