reverse the order in which sources are searched for in the package sets
This fixes an annoying discrepancy between the loading order and the
version control overriding order:

More specifically, autobuild files from a "generic" package set
might define convenience methods that are used with "more specific"
package sets.

But, then, we had to load them first (put them first in the list), which
means that the most specific package sets would *not* be able to override
the source definitions ...

So, now, the list of package set has to be written from the most generic
to the least generic, and everything follows.