we must first explicitely call #import on all packages to make automatic dependency discovery possible
The issue is that the package's #import method is in charge of
specifying what the package provides (as for instance pkg-config
packages). Therefore, if it is not called, it will not be possible
for selected packages to auto-discover what they depend on.

Note that the weakness of this method is that it works only if
all the packages are already checked out