support relative URLs in the bootstrapping command line
The issue was that the relative URL was passed on in the configuration
update code, which was trying to resolve it relatively to
Autoproj.root_dir. This is not possible, as we don't have an autoproj
installation yet. Idem for switch_config, as autoproj/ might be
deleted when we get there.

So, resolve relative directories as soon as possible in bootstrap
and switch_config