fix bootstrap and switch-config
Both had numerous bugs related to the refactoring of the
bin/autoproj tool, and the discovery that relative URLs were
not handled properly.

Added a new functionality: if the repo type and URL are not changed,
they don't have to be specified on the switch-config command line, i.e.

  autoproj switch-config branch=other_branch

is the same as

  autoproj switch-config <current VCS> <current URL> branch=other_branch