unify package and os dependencies
Starting from now, the calls to #depends_on, #depends_on_os_package
and the related fields in the manifest.xml file are equivalent.

It allows to use autoproj powerful building mechanisms to build
packages on some OSes which do not provide a prepackaged version,
while other OSes have a prepackaged version.

For instance, opencv 2.0 is available as a .deb on Debian
but not yet available on Ubuntu