since we can't check if OS dependencies are installed or not, do enable them on build
The issue is that do not know if a particular OS package is installed or not,
and we therefore cannot install it when not installed but avoiding updates.

On "non-bleeding-edge" OSes (like Ubuntu), it will not matter as packages
are updated only on new OS release, or for good reasons.

Bleeding-edge-OSes users (like Debian unstable) will have to deal with it, but
that's what you get when you use these.