unconditionally update the env.sh at the end of a build/doc/update
env.sh is updated at two steps: prepare and install. However, the
"install" part should be minimal (mostly calls to update_environment),
as #install -- unlike #prepare -- is called only if the install is
required.

Thus, at the end of a run, the exported env.sh always contains the
most up-to-date information on the currently built and installed
packages. There is therefore no reason whatsoever to not update it
(i.e. no information can be lost by not updating it)