improve behaviour for manifests that list only a few leaf packages
autoproj now has a sensible behaviour if only a few packages
(not package sets) are listed in the manifest. It was already
properly handling a complete build, but the various package
selection options would not work on the command line for packages
that are part of the build (because they are depended-upon), but are
not listed in the manifest.