do not update the same package twice
When we invoke the import and prepare targets, the dependencies import and
prepare are automatically called. However, there is two possible paths: either
the root package was selected, in which case the update has taken place, or not.

If a selected package did *not* get updated, reenable both the import and
prepare target, so that it gets updated.