fix partial package selections
The issue that this commit fixes is that #prepare could be called
multiple times, leading to having #build and/or #install called
multiple times as well.

This is fixed by realizing that #prepare does *not* need to be called
on packages that won't be built. The main reason is that #prepare
is meant to add the package dependencies, which we don't care about
if the package does not need to be built.