reword the package import logic (one more time)
Rationale: we actually don't need to traverse the whole
dependency graph to find the selected packages. We just have
to start *at* the selected packages and follow the trail
from there.

Another change: we first import all queued packages and then
prepare them. It makes the likelihood that automatic
dependency are found higher. However, the use of automatic dependencies
will be discouraged starting from the 1.5.0 release, as they
don't work in quite a few cases.