fix osdeps vs. autobuild package selection
This is an issue with the unified the handling of OS and
regular packages. Namely, the autobuild package was implicitely
selected when the layout contains package set names.

Instead, install OS deps if present. However, add an override
mechanism: the autobuild package will always be used instead
of the OS package if the package is listed in the manifest.