be more eager in env.sh updates
We now update on failure as well. The issue is that a failed initial build is
leading to an unusable partial installation.

It is safe to update on failed builds, as the packages are supposed to update
the environment information at both import and install time. Therefore, an
already built package will get registered, a new package whose build was
successful will also get registered. Only packages whose build failed will get
rejected.