fix autoproj build <dir>
If <dir> was referring to a package that was not directly selected
in the layout, it would *not* select it unless <dir> == pkg.srcdir

Instead, allow <dir> to select any package that is checked out. We
still avoid checking out packages that are loosely selected though.
I.e., to check out a package, one has to provide its package name