allow overriding the osdeps detection, and don't endlessly autodetect unknown OSes
The AUTOPROJ_OS environment variable can be used to force the osdeps
detection to a particular value.

Moreover, it can be set to empty, in which case the detection is disabled.
In that case, the OS is handled as if it was unknown. Moreover, this patch
fixes the problem that, when run on an unknown OS, autoproj would try to
autodetect the OS over and over again.