add support for metapackages
The goal of this commit is to separate the usage of "package set
names" from the actual list a package set defines. This will allow,
for instance, to not build a given package by default when a package
set is listed in the layout.

In other words, in:

  layout:
     - rock

"rock" does not refer to the package set directly, but to a metapackage
(a list of packages) which, by default, is initialized to all packages
that the "rock" package set defines.