fix unnecessary OS autodetection issue once and for all
The main issue with OS autodetection was that, in some modes,
we were forcing autodetection "afresh" by only resetting the
according option

It depends heavily on the ordering between calls to
OSDependencies.operating_system and the reset call. This led to
unnecessary OS redetections because the OS was not saved in the
config file.

This fixes it once and for all by forcing an autodetection and
saving the result in the config file, instead of simply resetting
the value