added a very minimal query language
This is meant to help developping shell scripts based on autoproj
itself. For instance, select all packages that are using cmake and
git is done with

  vcs=git:class=cmake

In the future, autoproj and autolocate are going to use this query
language. For now, keep them as they are until the system is properly
tested