bin: add autoproj create-set <pkg_set_name> packages
It creates a new package set and populates autobuild and source.yml
files with the packages given on the command line.

If packages on the command line depend on other packages from
other package sets, or are themselves defined in other package
sets, then these are listed in the imports.