properly handle packages that are defined inside an ignored package set but then removed from the corresponding metapackage
This fixes the following definition

in pkg.set:

   cmake_package 'my/package'
   remove_from_default 'my/package'

where pkg.set is ignored.

The current behaviour was that my/package was (properly) not marked
as ignored, but since it was defined with :dummy, it was doing nothing.

This commit uses the new #disable call to disable the package without
redefining it. It makes it possible, in particular, to inspect
ignored packages (since they are fully configured)