osdeps: fix handling of removed but not purged packages in AptDpkgManager
The bug is only showing if a package gets removed but not purged. The issue was
that the list of current packages were not removed in-between the package
blocks, which led the purged package to be marked as installed because the NEXT
package was installed. Removed packages are removed from the status file, and
are therefore handled properly