fix a long-standing issue with some packages disappearing from the environment
The issue was that we are loading the environment for all selected
packages, but are only loading the package manifests for the selected
packages. Some dependencies were missing, and therefore the list
of selected packages was wrong.

Eagerly load all package manifests for all defined packages, but
only warn when they have errors. Only the selected packages will cause
an error.