run 'gem pristine' only for the gems we know and use
gem pristine --all --extensions is broken for gems that are installed
via a prebuilt binary package that contain a C extension. In this case
rubygem tries to build the extension and of course fails as it does
not have the rights to the target directory.

This commit modifies autoproj to run pristine only on the packages we
know about (i.e. the rubygems we depend on). It modifies the mechanisms
so that other os-indep managers (e.g. pip) can provide the same service.