implement auto-save of the commit state in import_packages
This saves the commit state in the reflog of a ref called refs/autoproj.
The idea is taken from the stash implementation. It ensures that git gc
will expire the refs after a (configurable) amount of time, that the
log cannot be pushed by mistake, and that each log entry does not have
to have its predecessor as parent (it will always have the current
main configuration commit as parent instead)