autobuild: run common doc/test setup before we yield
There is a chicken-and-egg problem here. That is, if we want the block
to test whether tests are available for a package (to e.g. tune cmake
to build the tests), we need to run the common cmake setup so that
tests are enabled. Given that it is possible to override the tests/doc
targets in the blocks nowadays, it is harmless to packages that want
to use non-default test/doc targets.