fix os-detection for debian
On my Debian Jessie, the os-auto-detection returned [debian, 8.0], which
is wrong in the way that most od the rock-osdeps use the names "jessie"
or "sid" -- not the release-number.

this commit copies the debian-relevant logic out of
"guess_operating_system()". this allows using the more reliable
"os_from_os_release()" also for debian and just overwriting the
"version" array if unstable was detected.

Signed-off-by: Martin Zenzes <martin.zenzes@dfki.de>