make the mapping from command line selection to packages weak
We assume that the user knows what he is doing. Most of the time,
he will put a package name and that will fail if e.g. the package
is excluded. However, if he uses a catch-all (like orogen/ to build
all orogen packages), he obviously does not want it to fail because
some orogen packages are excluded