cleanup handling of exclusions/ignores/osdep overrides
This handling, as well as error generation, was spread all over the
place. This led to really hard to understand behaviour - especially
since some old code did not have the same semantic than new one.

This commit removes all non-straightforward error handling for
dependencies, apart from "does not exist", from the package
representation code. The handling of existent-but-not-available
OS dependencies is done during package import/prepare (in Ops::Import)
and
The CLI code is now responsible for all that stuff.

In addition, this opened the way to handle osdeps more uniformly.
OSDeps are now properly resolved when given on the command line -
in a way that's uniform with the source packages - and can be
ignored/excluded in the same way