remove calls of post install blocks during setup
There is no way to know whether a package got installed or not, only
whether it is imported or not. So, calling post_install can really only
be done by autobuild after #install.

The use of it I had (setup that depended on the presence of the package
dependencies) can for now be done with

   op = lambda do
      [setup code]
   end

   begin op.call
   rescue # dependencies are not installed
      pkg.post_install { op.call }
   end

If it becomes a common patter, maybe we should add it to the API.