ops: in Import, gracefully handle the Autobuild::InteractionRequired exception
Interaction is not allowed in parallel build. For some importers, one
can set the interactive flag in advance. For some others (e.g. the
archive importer), it's not possible as it depends on the need
for the importer to ask questions to the user.

This commit handles the InteractionRequired exception, moving the importer
to the interactive_imports set when it's rescued and calling it again
later.