convert the bootstrap and gem handling to a bundler-based setup
The major change is that bundler and autoproj can be privately installed
in resp. ".autoproj/bundler" and ".autoproj/autoproj" while the rest of
the gems are installed in the ".autoproj/gems" folder. *Or* we can
safely use a shared gem folder, bundler taking care of the versioning.

Bundler provides us with a much smoother experience workspace-wise
since it takes care of invalid cross-dependencies, and allows us
to snapshot the gem versions (by adding files called Gemfile* to
autoproj/ or autoproj/overrides.d)