fix issues concerning the validation of AUTOPROJ_CURRENT_ROOT and Dir.pwd
The test was embedded in Workspace.from_dir, which makes no sense
(if one wants to initialize from a given directory, then let him).
Moreover, the logic was broken since the default search dir now points
to AUTOPROJ_CURRENT_ROOT and the test was comparing that with ....
the value of AUTOPROJ_CURRENT_ROOT. Fix the test by looking at a workspace
that contains Dir.pwd instead.

Moreover, in from_environment, raise an error if AUTOPROJ_CURRENT_ROOT
is set but to a directory that is not an autoproj install. This was
already failing, but much later.