install: run the gem program with the current ruby interpreter
This makes sure that we load the gems that are bundled with *that*
interpreter, regardless of the gem program we found.

Not doing so was a workaround in case 'gem' is a shell shim (e.g.
rbenv), but that's not needed since we started resolving the gem
program from CONFIG['bindir'] directly, instead of PATH