bundler: really merge all the Gemfile snippets we get
Bundler does not accept having multiple gem specifications for the same
gem, but that's the intended use of the gemfile snippets. So, use
the Bundler API to load the Gemfile and merge them.

This allows to override a gem by a local source by adding it to
a local Gemfile in autoproj/Gemfile or autoproj/overrides.d/*.gemfile