install: install bundler with a '/usr/bin/env ruby' shebang
We've elected to install bundler only once for all ruby versions,
something bundler itself is very adapted to -- it's compatible with
everything.

However, by installing it without --env-shebang, when started from
the command line it would use the first ruby version that installed
it, ever. It is not a problem for autoproj, which specifies the full
path to the ruby executable, but is an issue when running bundler from
the command line.