install: rationalize the handling of specific gem dirs
Splitting the gem directories is mainly meaningless. We really cannot
guarantee anything as soon as we use a shared gem directory, as e.g.
that the shared gem directory for the workspace gems does not include
bundler, which would cause it to be used instead of the one installed
somewhere else (as would have been specified by --private-bundler)

So, either allow to provide a single shared gem location (in which
all of bundler, autoproj and workspace gems will be installed),
or to specify that the gems should be installed privately (in which
case there is a split between bundler/autoproj in the source workspace,
and the workspace gems in the prefix), but since both locations are
under control of that particular workspace, it sounds reasonable.

In addition, the layout of the bundler and autoproj gems has also
been changed. Bundler is installed in a platform/version subdirectory
(as it would for instance be in $HOME/.gems), and in the private
case, both bundler and autoproj are installed in .autoproj/gems

This change will cause early users to get a message telling them to
re-run autoproj_install.