define 'strict?' package managers
Strict managers maintain and enforce a list of packages that are
made available to the workspace. Example of such a manager is Bundler.
install() requires to be passed all the packages the workspace needs

The non-strict ones (e.g. APT) would make the packages, once installed,
available to the workspace. In their case install() requires only
the packages-to-be-installed at this level.