accept package names in Manifest#importer_definition_for and #overrides_for
This is a functionality used in the Rock package sets. Whenever an attempt to
define a package is detected, the Rock autobuild code might try to resolve the
package's branch to check whether it's part of the current flavor or not.

It requires to resolve even without having the package in the manifest (i.e.
disable name checks)