resolve relative paths early in package set resolution
The late-resolution that was done in VCSImporter#create_autobuild_importer
forced access to Autoproj.root_dir (obviously). This ensures that we use
the package-set-local root dir as base for resolution.

In general, we could also decide to forbid relative paths and force
the usage of $AUTOPROJ_ROOT in relative URLs.