ops: reset RUBYLIB in the autoproj shims
The bundler manager installs autoproj in both the core autoproj
install (.autoproj) and the general gem location (install/gems).
The latter might get out-of-sync when the former is updated.
This can actually lead to very weird situations.

Since env.sh contains the explicit RUBYLIB, running the current
shims would load the old version (and if that old version is buggy,
lead to a deadlock). This commit explicitely resets RUBYLIB to make
sure that the shims get a clean bundler environment from .autoproj/

This requires to have a specific system to install plugins for autoproj, as
the plugins now really need to be enabled by .autoproj/Gemfile