refactor the handling of ignore_errors
We do not leave the handling of ignore_errors to the individual
importers anymore. It was already the case in Ops::Import, this
commit makes it also the case in Ops::Configuration.

Both opts classes will aggregate the errors they are seeing into
a composite ImportFailed error, which is then