move configuration loading out of Workspace into Ops::Configuration
Ops::Configuration is needed to setup the package set list, so use
it as well to do the actual loading.

It changes all the 'ignore_errors' options to 'keep_going', which is
a lot more informative to what this option actually does.

It also fixes a bug in "keep going" mode, in which the configuration was
not loaded anymore and therefore we ended up with an empty configuration.