fix OSPackageResolver#merge when osdep entries that use recursive resolution collide
To avoid spurious warning, we're resolving the entries with
resolve_package, so that we can check that there is an override
for *this* particular OS.

The call to resolve_package would fail if the colliding entry has
a recursive entry that is not served in the same osdep file, given
that we've not merged yet.

Fix that by optionally avoiding recursive resolution.