make Package#depends_on resolve defined-but-unavailable packages as odeps (fixes #125)
Since all setup blocks are called, we really need #depends_on to
delay availability checks. Otherwise, an unused package for which
an osdep is defined but not present will make load fail in the
setup block.

The actual error of using an unavailable package is handled at
package-load time, which also handles optional dependencies
properly.