cli: fix environment export under error in 'update'
The main issue here is to make sure that the default list of
packages is available when we attempt to export env.sh.

Multiple problems:
 - if the package sets failed to update and keep_going is false,
   the load was aborted and some packages or package sets are
   simply undefined
 - if the package update is aborted, the package manifests are not
   all loaded

Go back to exporting the env.sh only if there is either no errors,
or if keep_going is set. The only change is that we do export even
if no packages get imported - to for instance reflect changes within
the package sets.