implement auto-exclusion of broken packages during update and build
So far, autoproj will fail hard if a package cannot be checked out,
nor if its manifest.xml has error.

This is a problem for CI environments for meta-targets (i.e. 'rock.all').

Allow passing --auto-exclude to exclude packages in these cases,
allowing the rest of the build to continue. Note that targets that
do not have weak dependencies will still fail (i.e. it is safe to enable
in all CI environments if important targets are marked as non-weak)