cli: make 'watch' robust to errors in the configuration
First of all, this changes the way notifiers are set up, to watch
everything in .autoproj/remotes and autoproj/. I don't see the
point in selecting files, which probably guarantees that we're
fatally going to forget one. Moreover, the scheme we had
was failing on boostrap, as the package sets are not present
at this point (and we were not trying to find them).

Then, in case updating the workspace fails, we try to instead
discover the active packages using the installation manifest.
If this fails, we just assume that the configuration will have to
be fixed by the user first.