test: fix update_autoproj test in workspace after changes in dependencies
The current test was fragile when we change dependencies, as it would
fail to get the local autoproj code as the main package. This would
actually also make the test useless, as we would test the last released
autoproj version, not the next.

Try to make sure this doesn't happen by providing absurd version numbers
for before *and* after the update (it was only after beforehand)