require setting an explicit flag to load ROS' package.xml
The problem here is that we have packages that do NOT have an
autoproj-compatible manifest.xml but do have a package.xml. Moreover,
there also have non-ROS, non-Rock build systems that use package.xml
(albeit with a different format). In order to avoid breaking existing
systems, require that a flag is explicitely set to load package.xml