fix some more corner-cases related to moving bundler in .local/
The main issue is that Gem would pick up a bundler already installed
in $HOME/.gem, which would work until we get to our shims, which
reset GEM_HOME and GEM_PATH to ensure an isolation of the Autoproj
workspace gems.

Make sure the bundler we find is indeed in #gems_gem_home, and
that it gets installed there regardless of the state of the rest
of the system.