explicitly set RUBYLIB to our bundler's lib/ folder when discovery rubylib itself
We were so far relying on the rubygems loading mechanism.
Unfortunately, Debian (and Ubuntu's) way to package gems
put them in the system rubylib instead of in the gem path.

This therefore caused autoproj to load the bundler from the
Ubuntu package instead of our gem-installed bundler. It started
to fail recently, since 1.x is packaged but autoproj is installing
2.x.