fix utility-related dependencies not being installed on first checkout
The utility enabled? flag is dependent on the utility availability,
which in turn is decided by testing for the presence of a test/ folder
for e.g. cmake packages.

The first mistake was that this test was done in the main context,
not a post_import block.

Then, post import blocks were called after the dependencies from
the manifest are applied to the package. Any change of utility
within the post import block would therefore not be taken into
consideration when interpreting the manifest file.