implement ignoring optional dependencies by default
The default policy (optional dependencies are added unless they are
excluded) is great in a development environment. However, for a more
embedded setup, the opposite policy (ignore optional dependencies
unless explicitely added) makes more sense.

Implement it. The flag is in the manifest, as it affects what gets
into the manifest directly. Note that it can also be changed
programatically through the API once the manifest is loaded.