fix(test): fix tests that install autoproj itself
These tests are rather complicated since we have to make sure we
don't interfere with the local system, as well as not getting
interference from it.

These changes ensure that we don't use the local autoproj gems in
all cases. I started having test failures on my dev machine because
of this, since I am using a rbenv-built ruby 2.7.1 for Rock and C
extensions don't get loaded by the Autoproj tests (that were using
the stock 2.7.0)

We also adds calls to with_unbundled_env where they were needed but
missing

Additionally, we remove the usage of a local gem server, that was
required for most installation tests pre-bundler, but aren't anymore
since bundler allows us to refer to the local dev checkout during
installation. The local gem server is kept only for the upgrade test.