fix: gem install location on Debian 11
Debian is (in)famous for monkey-patching rubygems to make it follow
their standard. In this case, it moves the default gems install location
to the XDG data dir.

It defines Gem.data_home to indicate where things are going to be
installed. Use its presence as an indication that we're on Debian
and workaround the problem.